added quantity
