+++ app/src/main/res/drawable/bg_quantity_badge.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="#FF4081"/>
+    <corners android:radius="50dp"/>
+</shape>
