+++ app/src/main/res/layout/item_inventory.xml
@@ -60,14 +60,39 @@
             android:text="Date Added: Not set"
             android:textColor="@android:color/darker_gray"
             android:textSize="14sp" />
+
+
     </LinearLayout>
 
-    <!-- Product Image on the right -->
-    <ImageView
-        android:id="@+id/productImage"
-        android:layout_width="94dp"
-        android:layout_height="84dp"
-        android:layout_gravity="center_vertical"
-        android:contentDescription="Product image"
-        android:src="@drawable/placeholder_image" />
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:gravity="center_horizontal">
+
+        <TextView
+            android:id="@+id/quantityBadge"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="4dp"
+            android:background="@drawable/bg_quantity_badge"
+            android:paddingStart="8dp"
+            android:paddingEnd="8dp"
+            android:paddingTop="2dp"
+            android:paddingBottom="2dp"
+            android:text="Qty: 1"
+            android:textColor="@android:color/white"
+            android:textSize="12sp"
+            android:textStyle="bold"
+            android:visibility="gone" />
+
+        <ImageView
+            android:id="@+id/productImage"
+            android:layout_width="94dp"
+            android:layout_height="84dp"
+            android:layout_gravity="center"
+            android:contentDescription="Product image"
+            android:src="@drawable/placeholder_image" />
+    </LinearLayout>
+
 </LinearLayout>
